Move service logic to timetravel endpoint 🎨
